+++ .devcontainer/devcontainer.json
@@ -1,7 +1,7 @@
 // For format details, see https://aka.ms/devcontainer.json.
 {
   "name": "Codefresh Support Package",
-  "image": "denoland/deno:2.0.3",
+  "image": "denoland/deno:2.0.6",
   "onCreateCommand": "apt-get update && apt-get install git zip -y",
   "customizations": {
     "vscode": {
