+++ ci/codefresh.yaml
@@ -1,4 +1,4 @@
-version: '1.0'
+version: "1.0"
 
 stages:
   - Clone
@@ -20,7 +20,7 @@ steps:
     title: Compiling
     stage: Build
     arguments:
-      image: denoland/deno:alpine-2.0.3
+      image: denoland/deno:alpine-2.0.6
       commands:
         - cf_export VERSION=$(cat VERSION)
         - deno task compile
